minimise scrolls to top of credibility signals card
