+++ src/components/NavItems/Assistant/AssistantScrapeResults/TextFooter.jsx
@@ -17,6 +17,8 @@ import { Translate } from "@Shared/Utils/Translate";
 import { getLanguageName } from "@Shared/Utils/languageUtils";
 import { i18nLoadNamespace } from "components/Shared/Languages/i18nLoadNamespace";
 
+import { scrollToElement } from "./assistantUtils";
+
 export default function TextFooter({
   classes,
   setDisplayOrigLang,
@@ -172,6 +174,7 @@ export function ExpandMinimise({
       fullWidth
       onClick={() => {
         setExpanded(!expanded);
+        expanded ? scrollToElement("credibility-signals", 100) : null;
       }}
     >
       {expanded ? (
@@ -191,7 +194,10 @@ export function ExpandMinimise({
     </Button>
   ) : (
     <IconButton
-      onClick={() => setExpanded(!expanded)}
+      onClick={() => {
+        setExpanded(!expanded);
+        expanded ? scrollToElement("credibility-signals", 100) : null;
+      }}
       sx={{
         "&:hover": {
           backgroundColor: "inherit",
