+++ src/components/NavItems/Assistant/AssistantScrapeResults/assistantUtils.jsx
@@ -515,3 +515,12 @@ export function ColourGradientScale({
     </Grid>
   );
 }
+
+export function scrollToElement(id, padding = 0) {
+  const element = document.getElementById(id);
+  if (element) {
+    const targetPosition =
+      element.getBoundingClientRect().top + window.scrollY - padding;
+    window.scrollTo({ top: targetPosition, behavior: "smooth" });
+  }
+}
